+++ pyecf/utils/gcd.py
