+++ pyecf/utils/misc.py
